requirements.txt
geodaisy
matplotlib
numpy
Pillow
pandas
shapely
tqdm

